Attempt to find supercollider boot file before default falling back
This updates the vim plugin to support custom supercollider boot files.
If `tidal_sc_enable` is `1`, the plugin will first tranverse through
parent directories to find either a `boot.sc` or `boot.scd` file. If one
is found, this file is used with the `sclang` command. If not, the
`tidal_sc_boot_fallback` file is used if specified.

This commit also changes the `superdirt` references within the plugin to
`sc` or `supercollider` in order to show that the support for this boot
file is not necessarily superdirt specific.